requirements.txt
pyTelegramBotAPI==4.9.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1